+++ mycodo/utils/functions.py
@@ -92,6 +92,7 @@ def parse_function_information(exclude_custom=False):
             dict_controllers = dict_has_value(dict_controllers, function_custom, 'options_disabled')
             dict_controllers = dict_has_value(dict_controllers, function_custom, 'dependencies_module')
             dict_controllers = dict_has_value(dict_controllers, function_custom, 'dependencies_message')
+            dict_controllers = dict_has_value(dict_controllers, function_custom, 'function_actions')
             dict_controllers = dict_has_value(dict_controllers, function_custom, 'custom_options')
             dict_controllers = dict_has_value(dict_controllers, function_custom, 'custom_channel_options')
             dict_controllers = dict_has_value(dict_controllers, function_custom, 'custom_actions')